meetup/requirements.txt
Django==1.10.2
factory-boy==2.7.0
fake-factory==0.7.2
psycopg2==2.6.2
python-dateutil==2.5.3
django-debug-toolbar==1.6
